Wip ordering

+++ classes/ListScreenRepository/Storage.php
@@ -4,6 +4,7 @@ namespace AC\ListScreenRepository;
 
 use AC\ListScreen;
 use AC\ListScreenCollection;
+use MongoDB\Driver\Exception\LogicException;
 
 final class Storage {
 
@@ -49,11 +50,26 @@ final class Storage {
 	}
 
 	/**
-	 * Returns the repositories last in, first out
+	 * Returns the repositories last in, first out and writable + no rules first
+	 *
 	 * @return Storage\ListScreenRepository[]
 	 */
-	private function get_repositories_lifo() {
-		return array_reverse( $this->repositories );
+	private function get_repositories_ordered() {
+		$has_rules = [];
+		$is_writable = [];
+		$is_readable = [];
+
+		foreach ( array_reverse( $this->repositories ) as $repository ) {
+			if ( $repository->has_rules() ) {
+				$has_rules[] = $repository;
+			} elseif ( $repository->is_writable() ) {
+				$is_writable[] = $repository;
+			} else {
+				$is_readable[] = $repository;
+			}
+		}
+
+		return array_merge( $is_readable, $is_writable, $has_rules );
 	}
 
 	/**
@@ -66,7 +82,7 @@ final class Storage {
 	public function find_all( array $args = [], Filter $filtering = null, Sort $sorting = null ) {
 		$list_screens = new ListScreenCollection();
 
-		foreach ( $this->get_repositories_lifo() as $repository ) {
+		foreach ( $this->get_repositories_ordered() as $repository ) {
 			foreach ( $repository->find_all( $args ) as $list_screen ) {
 				if ( ! $list_screens->contains( $list_screen ) ) {
 					$list_screens->add( $list_screen );
@@ -91,7 +107,7 @@ final class Storage {
 	 * @return ListScreen|null
 	 */
 	public function find( $id ) {
-		foreach ( $this->get_repositories_lifo() as $repository ) {
+		foreach ( $this->get_repositories_ordered() as $repository ) {
 			if ( ! $repository->exists( $id ) ) {
 				continue;
 			}
@@ -113,21 +129,30 @@ final class Storage {
 	}
 
 	public function save( ListScreen $list_screen ) {
-		foreach ( $this->get_repositories_lifo() as $repository ) {
-			$match = ! $repository->has_rules() || $repository->get_rules()->match( $list_screen );
-
-			if ( $match && $repository->is_writable() ) {
-				$repository->save( $list_screen );
-
-				return;
-			}
-		}
+		$this->update( $list_screen, 'save' );
 	}
 
 	public function delete( ListScreen $list_screen ) {
-		foreach ( $this->get_repositories_lifo() as $repository ) {
-			if ( $repository->is_writable() ) {
-				$repository->delete( $list_screen );
+		$this->update( $list_screen, 'delete' );
+	}
+
+	private function update( ListScreen $list_screen, $action ) {
+		foreach ( $this->get_repositories_ordered() as $repository ) {
+			$match = ! $repository->has_rules() || $repository->get_rules()->match( $list_screen );
+
+			if ( $match && $repository->is_writable() ) {
+				switch ( $action ) {
+					case 'save':
+						$repository->save( $list_screen );
+
+						break;
+					case 'delete':
+						$repository->delete( $list_screen );
+
+						break;
+					default:
+						throw new LogicException( 'Invalid action for update call.' );
+				}
 
 				return;
 			}
